+++ app/src/navigation/AppTabs.tsx
@@ -58,6 +58,8 @@ import MyRoutineScreen from "../screens/routine/MyRoutineScreen";
 import GroupRoutineScreen from "../screens/routine/GroupRoutineScreen";
 import RoutineEditorScreen from "../screens/routine/RoutineEditorScreen";
 import CoverManageScreen from "../screens/routine/CoverManageScreen";
+import DailyNoteScreen from "../screens/routine/DailyNoteScreen";
+import BellScheduleScreen from "../screens/routine/BellScheduleScreen";
 import SectionPickerScreen from "../screens/common/SectionPickerScreen";
 import AdminHomeScreen from "../screens/admin/AdminHomeScreen";
 import ImportScreen from "../screens/admin/ImportScreen";
@@ -192,6 +194,8 @@ function RoutineNavigator(): React.ReactElement {
       <RoutineStack.Screen name="GroupRoutine" component={GroupRoutineScreen} options={{ title: STR.groupRoutineTitle }} />
       <RoutineStack.Screen name="RoutineEditor" component={RoutineEditorScreen} options={{ title: STR.editRoutineTitle }} />
       <RoutineStack.Screen name="CoverManage" component={CoverManageScreen} options={{ title: STR.coverManageTitle }} />
+      <RoutineStack.Screen name="DailyNote" component={DailyNoteScreen} options={{ title: STR.dailyNoteTitle }} />
+      <RoutineStack.Screen name="BellSchedule" component={BellScheduleScreen} options={{ title: STR.bellScheduleTitle }} />
       <RoutineStack.Screen name="SectionPicker" component={SectionPickerScreen} options={{ title: STR.pickSection }} />
     </RoutineStack.Navigator>
   );